Added Meshes and Jaw animation
Need to figure out how to implement the iris of the eyes and how to have the skull set up with children. Currently it is drawing two sets of the bottom teeth with how I have things implemented. Will need to figure out the best approach for hierarchy, probably going to use: top (1) - Eye1 (2) Eye2 (2) Jaw (2)* Upper Teeth (2) - Lower Teeth (3)